Quick update - corrected channel-messages to be in the right order, tho without timestamp for the time being. Will add if I have the time.
So, messages work the way it's intended, ish. Guest can send messages in the open chat but not to other users, users can send to users and in the open channel, as well as remove themselves. Admin can do all of the above and remove others.

Moving on to make sure when the guest "logs out" it self-deletes and if the user logs out it saves the login info, same with admin.