Finished up some troubles with guest-login and admin-login - hard to tell why they clashed but now they're working.
Everything is working as intended - guests cannot write dms or in closed channels, users can as well as create and delete channels, admins can do all that as well as delete users.

So yeah, seems to be working now - code might be a bit messy though so I will not be touching it anymore right now.